% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{Attribute}
\alias{Attribute}
\alias{CategoricalAttribute}
\title{Attribute specification}
\usage{
Attribute(
  dist_fn,
  distort_prob_prior,
  distort_dist_prior = DirichletProcess(Inf),
  exclude_entity_value = TRUE,
  entity_dist_prior = NULL
)

CategoricalAttribute(
  distort_prob_prior,
  distort_dist_prior = DirichletProcess(Inf),
  exclude_entity_value = TRUE,
  entity_dist_prior = NULL
)
}
\arguments{
\item{dist_fn}{A vectorized distance function that takes two vectors of
attribute values \eqn{y} and \eqn{x} as arguments, and returns a vector
containing the elementwise distances between the values. The function
is not assumed to be symmetric in it's arguments. The first argument
\eqn{y} corresponds to a (latent) entity attribute value and the second
argument \eqn{x} corresponds to an (observed) record attribute value.
This function parameterizes the distortion distribution (see below).
The efficiency/scalability of inference can be improved by returning a
distance of \code{Inf} whenever \eqn{x} is unlikely to be a distortion of
\eqn{y}. This can be achieved by wrapping a distance function in
\code{\link{transform_dist_fn}}.}

\item{distort_prob_prior}{A \code{\link{BetaRV}} object. Specifies the prior
on the distortion probability for the attribute.}

\item{distort_dist_prior}{A \code{\link{DirichletProcess}} object. Specifies
the prior on the distortion distribution for a record value \eqn{x}
with entity value \eqn{y}. A Dirichlet Process with a small
concentration parameter yields a more concentrated distribution, which
may be useful for large attribute domains. If the concentration parameter
is set to \code{Inf}, the Dirichlet Process collapses to the base distribution
(i.e. the prior is a point mass on the base distribution).
The base distribution is set to \eqn{p(x|y) \propto \exp(-\mathrm{dist_fn}(y,x))}{p(x|y) ∝ exp(-dist_fn(y,x))}. 
Defaults to \code{DirichletProcess(Inf)}.}

\item{exclude_entity_value}{A boolean specifying whether the entity
value \eqn{y} should be excluded from the support of the distortion
distribution for record value \eqn{x}. If TRUE, the distance returned
by \eqn{\mathrm{dist_fn}(y,y)}{dist_fn(y,y)} is replaced by \code{Inf} for all
values \eqn{y}. Defaults to TRUE.}

\item{entity_dist_prior}{A \code{\link{DirichletRV}} object, numeric vector or
NULL. This parameter specifies the prior on the entity attribute
distribution.

If a \code{\link{DirichletRV}} object is passed, a Dirichlet prior is placed on the
entity attribute distribution.

An arbitrary fixed distribution can be specified by passing a numeric
vector. It must have the same length as the domain of the attribute and
be normalized to 1.

If NULL, the entity attribute distribution is fixed to the empirical
distribution of the attribute values, as computed from the observed data.}
}
\value{
The \code{Attribute} constructor returns an \code{Attribute} object.

The \code{CategoricalAttribute} constructor returns a
\code{CategoricalAttribute} object.
It is intended for modeling categorical attributes, and uses a
a constant distance function.
}
\description{
Constructs an \code{Attribute} object, which stores model parameters associated
with an attribute (e.g. 'name of company', 'zip code', 'date of birth' etc.).
}
\details{
The model parameters specified by this function are:
\itemize{
\item The prior on the \emph{entity attribute distribution}, denoted
\code{entity_dist_prior} in the parameter list. The entity attribute distribution
is an important component of the entity model. When instantiating a new
entity, this distribution selects a value for the corresponding
attribute---e.g. "July 10 1973" for a "date of birth" attribute.
\item The prior on the \emph{distortion probability} for the attribute, denoted
\code{distort_prob_prior} in the parameter list. This controls the likelihood
that values of the attribute are distorted when they are recorded in the
data.
\item The prior on the \emph{distortion distribution} for the attribute, denoted
\code{distort_dist_prior} in the parameter list. The distortion distribution
is responsible for generating a distorted attribute value \eqn{x}
(recorded in the data) given the latent entity attribute value \eqn{y}.
It is parameterized by a distance function (see next bullet point).
\item The \emph{attribute distance function}, denoted \code{dist_fn} in the parameter
list. It takes arguments \eqn{y} and \eqn{x} and returns a distance,
which quantifies the likelihood that \eqn{x} is a distorted alternative
to \eqn{y}. The function need not satisfy the properties of a distance
metric.
}
}
\seealso{
A named list of \code{Attribute} objects is required when initializing the model
(see \code{\link{initialize_exer}}).
}
