% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttributeIndex.R
\name{AttributeIndex}
\alias{AttributeIndex}
\title{Attribute Index}
\usage{
AttributeIndex(values, dist_fn, exclude_entity_value, adjust_dist)
}
\arguments{
\item{values}{A vector of attribute values (e.g. a column from the source
data) represented as a factor.}

\item{dist_fn}{A vectorized distance function or NULL (for a constant
distance function).}

\item{exclude_entity_value}{A boolean specifying whether the entity
value \eqn{y} should be excluded from the support of the distortion
distribution for record value \eqn{x}.}

\item{adjust_dist}{Whether to adjust \code{dist_fn(y, x)} by subtracting
the log relative frequency of x in \code{values}.}
}
\value{
An \code{AttributeIndex} object.
}
\description{
Builds an \code{AttributeIndex} object for a collection of observed attribute
values.
}
\details{
The index includes quantities that are required for running inference
such as the empirical distribution over the domain and an index for
conducting range queries. This is an internal function and should not be
required unless constructing an \code{\linkS4class{ExchangERModel}} object manually.
}
\note{
This function supports parallel evaluation via the \code{future} package.
To configure the evaluation strategy, see \code{future::plan}.
}
\keyword{internal}
