% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaRV.R
\name{BetaRV}
\alias{BetaRV}
\title{Beta-distributed Random Variable}
\usage{
BetaRV(shape1, shape2)
}
\arguments{
\item{shape1, shape2}{Positive shape parameters of the Beta distribution}
}
\value{
The \code{BetaRV} constructor returns a \code{BetaRV} object,
which is a subclass of \code{\linkS4class{RV}}.
}
\description{
Represents a random variable with a Beta distribution.
}
\details{
The density of the Beta distribution with parameters \eqn{shape1 = a} and
\eqn{shape2 = b} is given by
\deqn{p(X = x) = \frac{\Gamma(a+b)}{\Gamma(a) \Gamma(b)} x^{a-1} (1-x)^{b-1}.}{p(X = x) = Gamma(a+b)/(Gamma(a) Gamma(b)) x^{a-1} (1-x)^{b-1}.}
for \eqn{0 < x < 1}, \eqn{a > 0} and \eqn{b > 0}.
}
\examples{
# A uniform distribution on the unit interval 
x <- BetaRV(1, 1)
mean(x)

}
\seealso{
Other random variables defined in this package
include \code{\link{GammaRV}}, \code{\link{DirichletRV}} and \code{\link{ShiftedNegBinomRV}}.
}
