% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoundedNBDRP.R
\name{BoundedNBDRP}
\alias{BoundedNBDRP}
\title{Bounded NBD Random Partition}
\usage{
BoundedNBDRP(shape1, shape2, n, alpha, a, q)
}
\arguments{
\item{shape1, shape2}{Shape parameters for the shifted beta-binomial base
distribution on the cluster sizes. Must be positive numeric scalars.}

\item{n}{Number of trials for the binomial base distribution on the cluster
sizes. Must be a non-negative integer scalar.}

\item{alpha}{Concentration parameter for the Dirichlet process prior on
the cluster sizes. Must be a positive numeric scalar.}

\item{a}{Target for number of successful trials for the shifted negative
binomial prior on the number of clusters. Must be a positive numeric
scalar.}

\item{q}{Probability of success for the shifted negative binomial prior on
the number of clusters. Must be a numeric scalar on the unit interval.}
}
\value{
The constructor \code{BoundedNBDRP} returns a \code{BoundedNBDRP} object which
is a subclass of \code{\linkS4class{RP}}.
}
\description{
TODO: update description.
A Kolchin partition model where the number of clusters is drawn from a
shifted negative binomial distribution, and the cluster sizes are drawn from
a distribution which is itself drawn from a Dirichlet process prior with
concentration \code{alpha} and a shifted beta-binomial base distribution.
Following the paramterization of the negative binomial distribution in
\code{\link[stats:NegBinomial]{stats::NegBinomial}}, the shifted negative binomial distribution for the
cluster sizes has parameters \eqn{`size` = a} and \eqn{`prob` = q}.
The shifted beta-binomial base distribution has parameters
\eqn{`shape1` = s_1} and \eqn{`shape2` = s_2} with probability mass function
\deqn{p(K = k) = {n \choose k - 1} \frac{\mathrm{B}(k - 1 + s_1, n - k + 1 + s_2)}{\mathrm{B}(s_1, s_2)}}
over the integers \eqn{k = 1, 2, \ldots, n + 1}.
}
\references{
Brenda Betancourt, Giacomo Zanella, Hannah Wallach, Jeffrey W. Miller, Abbas
Zaidi & Rebecca C. Steorts (2016): Flexible Models for Microclustering with
Application to Entity Resolution, \emph{Advances in Neural Information Processing Systems}, 29
URL: https://proceedings.neurips.cc/paper_files/paper/2016/file/670e8a43b246801ca1eaca97b3e19189-Paper.pdf
}
\seealso{
Other random partitions defined in this package include \code{\link{PitmanYorRP}} and
\code{\link{EwensRP}}.
}
