% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R, R/hello.R
\name{convert_to_eur}
\alias{convert_to_eur}
\title{EUR/USD converter}
\usage{
convert_to_eur(val)

convert_to_eur(val)
}
\arguments{
\item{val}{numeric vector}
}
\value{
value_eur numeric value

value_eur numeric value
}
\description{
EUR/USD converter

EUR/USD converter
}
