% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirichletProcess.R
\name{DirichletProcess}
\alias{DirichletProcess}
\title{Dirichlet Process}
\usage{
DirichletProcess(alpha, H = NULL)
}
\arguments{
\item{alpha}{Concentration parameter. Must be a strictly positive real
number or a \code{\link{GammaRV}} hyperprior.}

\item{H}{Base distribution. Currently only NULL values are supported.}
}
\value{
A \code{DirichletProcess} object.
}
\description{
Represents a Dirichlet process.
}
