% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirichletRV.R
\name{DirichletRV}
\alias{DirichletRV}
\title{Dirichlet-distributed Random Vector}
\usage{
DirichletRV(alpha)
}
\arguments{
\item{alpha}{Positive concentration parameter vector. If a scalar is
passed, this value is repeated for each entry in the vector.}
}
\value{
The \code{DirichletRV} constructor returns a \code{DirichletRV} object,
which is a subclass of \code{\linkS4class{RV}}.
}
\description{
Represents a random vector with a Dirichlet distribution.
}
\details{
With concentration parameter vector \eqn{alpha = a} of length \eqn{K},
the density is
\deqn{p(X = x) = \frac{\prod_{i=1}^{K} \Gamma(\alpha_i)}{\Gamma(\sum_{i=1}^{K} \alpha_i)} \prod_{i=1}^{K} x_i^{\alpha_i-1} \exp(- b x).}{p(X = x) = (prod_i Γ(ɑ_i))/(Γ(sum_i ɑ_i)) prod_i x_i^{ɑ_i-1} exp(- b x).}
for vector \eqn{x} in the \eqn{K - 1} simplex.
}
\examples{
x <- DirichletRV(c(1,1,1))
mean(x)

}
\seealso{
Other random variables defined in this package
include \code{\link{BetaRV}}, \code{\link{GammaRV}} and \code{\link{ShiftedNegBinomRV}}.
}
