% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RV.R, R/GammaRV.R, R/BetaRV.R,
%   R/ShiftedNegBinomRV.R
\name{drv}
\alias{drv}
\alias{drv,numeric,GammaRV-method}
\alias{drv,numeric,BetaRV-method}
\alias{drv,numeric,ShiftedNegBinomRV-method}
\title{Density Function}
\usage{
drv(x, rv, log, ...)

\S4method{drv}{numeric,GammaRV}(x, rv, log = FALSE, ...)

\S4method{drv}{numeric,BetaRV}(x, rv, log = FALSE, ...)

\S4method{drv}{numeric,ShiftedNegBinomRV}(x, rv, log = FALSE, ...)
}
\arguments{
\item{x}{a vector of quantiles}

\item{rv}{a \code{\linkS4class{RV}} object}

\item{log}{a logical; if TRUE, probabilities are given as log(p).}

\item{...}{other parameters passed on to other methods.}
}
\value{
Returns the density of the random variable at \code{x}.
}
\description{
Computes the density of a random variable at a point
}
\section{Methods (by class)}{
\itemize{
\item \code{x = numeric,rv = GammaRV}: Specialization for \code{\link{GammaRV}}

\item \code{x = numeric,rv = BetaRV}: Specialization for \code{\link{BetaRV}}

\item \code{x = numeric,rv = ShiftedNegBinomRV}: Specialization for \code{\link{ShiftedNegBinomRV}}
}}

\examples{
x <- BetaRV(1, 1)
sd(x)

}
