% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESCNBRP.R
\name{ESCNBRP}
\alias{ESCNBRP}
\title{ESC-NB Random Partition}
\usage{
ESCNBRP(p, r)
}
\arguments{
\item{p}{Probability of success. Must be on the unit interval \eqn{(0, 1]}
or a \code{\link{BetaRV}} object.}

\item{r}{Target number of successes before negative binomial experiment
is stopped. Must be strictly a positive real or a \code{\link{GammaRV}} object.}
}
\value{
The constructor \code{ESCNBRP} returns a \code{ESCNBRP} object which is a
subclass of \code{\linkS4class{RP}}.
}
\description{
Represents an ESC-NB (Exchangeable Sequences of Clusters-Negative Binomial)
random partition.
}
\details{
TODO
}
\references{
Brenda Betancourt, Giacomo Zanella & Rebecca C. Steorts (2020): Random
Partition Models for Microclustering Tasks, \emph{Journal of the American
Statistical Association}, DOI: 10.1080/01621459.2020.1841647.
}
\seealso{
Other random partitions defined in this package include \code{\link{PitmanYorRP}} and
\code{\link{EwensRP}}.
}
