% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R, R/hello.R
\name{eur}
\alias{eur}
\title{EUR pretty printer}
\usage{
eur(value)

eur(value)
}
\arguments{
\item{value}{numeric vector}
}
\value{
EUR character string

EUR character string
}
\description{
EUR pretty printer

EUR pretty printer
}
