% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PitmanYorRP.R
\name{EwensRP}
\alias{EwensRP}
\title{Ewens Random Partition}
\usage{
EwensRP(alpha)
}
\arguments{
\item{alpha}{Positive concentration parameter or a \code{\link{GammaRV}}.}
}
\value{
The constructor \code{EwensRP} returns a \code{PitmanYorRP}
object which is a subclass of \code{\linkS4class{RP}}.
}
\description{
Represents a Ewens random partition.
}
\details{
This is a special case of the Pitman-Yor partition (see \code{\link{PitmanYorRP}})
for which the discount parameter \eqn{d = 0}. It exhibits distinct
asymptotic behavior: the number of clusters grows logarithmically in
the number of data points.
}
\seealso{
Other random partitions defined in this package include \code{\link{PitmanYorRP}} and
\code{\link{GeneralizedCouponRP}}.
}
