% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchanger-package.R
\docType{package}
\name{exchanger-package}
\alias{exchanger}
\alias{exchanger-package}
\title{exchanger: Bayesian Entity Resolution with Exchangeable Random Partition Priors}
\description{
Implements a Bayesian model for entity resolution as described 
    in Marchant et al. (2020). The model assumes records are generated from 
    a latent population of entities via a distortion process. The records are 
    assumed to be (infinitely) exchangeable, and several priors on the 
    linkage/clustering structure compatible with this assumption are 
    implemented. The model is fitted using a Markov chain Monte Carlo 
    algorithm.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cleanzr/exchanger}
  \item Report bugs at \url{https://github.com/cleanzr/exchanger/issues}
}

}
\author{
\strong{Maintainer}: Neil Marchant \email{ngmarchant@gmail.com}

Authors:
\itemize{
  \item Rebecca Steorts \email{beka@stat.duke.edu} [thesis advisor]
  \item Benjamin Rubinstein \email{brubinstein@unimelb.edu.au} [thesis advisor]
}

}
\keyword{internal}
