% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_inference.R
\docType{class}
\name{ExchangERResult-class}
\alias{ExchangERResult-class}
\title{Class for MCMC output}
\description{
Class for MCMC output
}
\section{Slots}{

\describe{
\item{\code{history}}{a list containing parameters of interest and summary
statistics along the Markov chain.}

\item{\code{state}}{an \code{\linkS4class{ExchangERModel}} object which represents the state
of the model at the end of the Markov chain.}
}}

