% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exer-package.R
\docType{package}
\name{exer-package}
\alias{exer}
\alias{exer-package}
\title{exer: Bayesian Entity Resolution with Exchangeable Random Partition Priors}
\description{
This package implements a Bayesian model for entity resolution 
    as presented in Marchant et al. (2020). The model assumes records are 
    generated from a latent population of entities via a distortion process. 
    Entity resolution is performed by inferring the latent linkage structure 
    using a Markov chain Monte Carlo algorithm.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cleanzr/exer}
  \item Report bugs at \url{https://github.com/cleanzr/exer/issues}
}

}
\author{
\strong{Maintainer}: Neil Marchant \email{ngmarchant@gmail.com}

Authors:
\itemize{
  \item Rebecca Steorts \email{beka@stat.duke.edu} [thesis advisor]
  \item Benjamin Rubinstein \email{brubinstein@unimelb.edu.au} [thesis advisor]
}

}
\keyword{internal}
