% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,ExchangERFit-method}
\alias{extract,ExchangERFit-method}
\alias{extract}
\title{Extract Samples}
\usage{
\S4method{extract}{ExchangERFit}(x, params = NULL, include = TRUE)
}
\arguments{
\item{x}{an \code{\linkS4class{ExchangERFit}} object}

\item{params}{an optional character vector specifying the names of
parameters/summary statistics to extract. If not specified, all
parameters and summary statistics are extracted.}

\item{include}{a logical scalar indicating whether the parameters in
\code{params} should be included (TRUE) or excluded (FALSE).}
}
\value{
If params is of length 1, the samples for the requested parameter are
returned as a \code{\link[coda:mcmc]{coda::mcmc}} object.

If params is NULL or of length greater than 1, the parameters are returned
in a named list. The samples for each parameter are represented as a
\code{\link[coda:mcmc]{coda::mcmc}} object.
}
\description{
Extract samples from a fitted model
}
\examples{
## Initialize a model for entity resolution of RLdata500
library(comparator)
distort_prior <- BetaRV(1,5)
clust_prior <- PitmanYorRP(alpha = GammaRV(1.0, 1.0), d = BetaRV(1.0, 1.0))
attr_params <- list(
  fname_c1 = CategoricalAttribute(distort_prior),
  lname_c1 = CategoricalAttribute(distort_prior),
  bd = CategoricalAttribute(distort_prior),
  by = CategoricalAttribute(distort_prior),
  bm = CategoricalAttribute(distort_prior)
)
model <- exchanger(RLdata500, attr_params, clust_prior)

## Run inference
fit <- run_inference(model, n_samples=100, burnin_interval=100)

## Extract samples of the linkage structure
links <- extract(fit, "links")

}
