% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_inference.R
\name{extract}
\alias{extract}
\title{Extract Samples}
\usage{
extract(x, params = NULL, include = TRUE)
}
\arguments{
\item{x}{an \code{\linkS4class{ExchangERFit}} object}

\item{params}{an optional character vector specifying the names of
parameters/summary statistics to extract. If not specified, all
parameters and summary statistics are extracted.}

\item{include}{a logical scalar indicating whether the parameters in
\code{params} should be included (TRUE) or excluded (FALSE).}
}
\value{
If params is of length 1, the samples for the requested parameter are
returned as a \code{\link[coda:mcmc]{coda::mcmc}} object.

If params is NULL or of length greater than 1, the parameters are returned
in a named list. The samples for each parameter are represented as a
\code{\link[coda:mcmc]{coda::mcmc}} object.
}
\description{
Extract samples from a fitted model
}
\examples{
#' ## Initialize a model for ER of RLdata500
library(comparator)
distortionPrior <- BetaRV(1,5)
clust_prior <- PitmanYorRP(alpha = GammaRV(1.0, 1.0), d = BetaRV(1.0, 1.0))
attr_params <- list(
  "fname_c1" = Attribute(Levenshtein(), distortionPrior),
  "lname_c1" = Attribute(Levenshtein(), distortionPrior),
  "bd" = CategoricalAttribute(distortionPrior),
  "by" = CategoricalAttribute(distortionPrior),
  "bm" = CategoricalAttribute(distortionPrior)
)
model <- exchanger(RLdata500, attr_params, clust_prior)

## Run inference
fit <- run_inference(model, n_samples=2000, burnin_interval=1000)

## Extract samples of the linkage structure
links <- extract(fit, "links")

}
