% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaRV.R
\name{GammaRV}
\alias{GammaRV}
\title{Gamma Random Variable}
\usage{
GammaRV(shape, rate)
}
\arguments{
\item{shape}{Positive real-valued shape hyperparameter}

\item{rate}{Positive real-valued rate hyperparameter}
}
\value{
The \code{GammaRV} constructor returns a \code{GammaRV} object,
which is a subclass of \code{\linkS4class{RV}}.
}
\description{
Represents a random variable with a Gamma distribution.
}
\details{
\code{GammaRV} constructs a \code{GammaRV} object.
The parameterization used here is in terms of a shape parameter
\eqn{shape = a > 0} and rate parameter \eqn{rate = b > 0}.
The density is
\deqn{p(X = x) = \frac{b^a}{\Gamma(a)} x^{a-1} \exp(- b x).}{p(X = x) = (b^a)/(Γ(a)) x^{a-1} exp(- b x).}
for \eqn{x \geq 0}{x >= 0}.
}
\examples{
x <- GammaRV(1, 1)
mean(x)

}
\seealso{
Other random variables defined in this package
include \code{\link{BetaRV}}, \code{\link{DirichletRV}} and \code{\link{ShiftedNegBinomRV}}.
}
