% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralizedCouponRP.R
\name{GeneralizedCouponRP}
\alias{GeneralizedCouponRP}
\title{Generalized Coupon Random Partition}
\usage{
GeneralizedCouponRP(m, kappa)
}
\arguments{
\item{m}{Number of coupons. Must be a strictly positive integer
or a \code{\link{ShiftedNegBinomRV}} object.}

\item{kappa}{Concentration parameter. Must be strictly a positive real or
a \code{\link{GammaRV}} object.}
}
\value{
The constructor \code{GeneralizedCouponRP} returns a
\code{GeneralizedCouponRP} object which is a subclass of \code{\linkS4class{RP}}.
}
\description{
Represents a generalized coupon random partition.
}
\details{
Generalized coupon random partitions are a subset of Ewens-Pitman random
partitions for which \eqn{\kappa = - \sigma > 0}{κ = -σ > 0} and
\eqn{\alpha = m \kappa}{ɑ = m κ} for some natural number \eqn{m}.
These partitions can be generated by drawing coupons with replacement from
a finite population of size \eqn{m}, where the mixing proportions are
drawn from a symmetric Dirichlet distribution with concentration parameter
\eqn{\kappa}{κ}. The coupon-collector's partition is obtained in the limit
\eqn{\kappa \to \infty}{κ -> Inf}.
}
\seealso{
Other random partitions defined in this package include \code{\link{PitmanYorRP}} and
\code{\link{EwensRP}}.
}
