% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_inference.R
\docType{class}
\name{MCMCResult-class}
\alias{MCMCResult-class}
\title{Class for MCMC output}
\description{
Class for MCMC output
}
\section{Slots}{

\describe{
\item{\code{history}}{a list containing parameters of interest and summary
statistics along the Markov chain.}

\item{\code{state}}{an \code{\linkS4class{EXERModel}} object which represents the state
of the model at the end of the Markov chain.}
}}

