% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RV.R, R/GammaRV.R, R/BetaRV.R, R/DirichletRV.R,
%   R/ShiftedNegBinomRV.R
\name{mean}
\alias{mean}
\alias{mean,GammaRV-method}
\alias{mean,BetaRV-method}
\alias{mean,DirichletRV-method}
\alias{mean,ShiftedNegBinomRV-method}
\title{Mean of a Random Variable}
\usage{
mean(x, ...)

\S4method{mean}{GammaRV}(x)

\S4method{mean}{BetaRV}(x)

\S4method{mean}{DirichletRV}(x)

\S4method{mean}{ShiftedNegBinomRV}(x)
}
\arguments{
\item{x}{a \code{\linkS4class{RV}} object}

\item{...}{other parameters passed on to other methods.}
}
\value{
Returns the mean of the random variable
}
\description{
Computes the mean (expected value) of a random variable
}
\section{Methods (by class)}{
\itemize{
\item \code{GammaRV}: Specialization for \code{\link{GammaRV}}

\item \code{BetaRV}: Specialization for \code{\link{BetaRV}}

\item \code{DirichletRV}: Specialization for \code{\link{DirichletRV}}

\item \code{ShiftedNegBinomRV}: Specialization for \code{\link{ShiftedNegBinomRV}}
}}

\examples{
x <- BetaRV(1, 1)
mean(x)

}
