% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{mp_clusters}
\alias{mp_clusters}
\alias{mp_clusters.matrix}
\alias{mp_clusters.ExchangERResult}
\alias{mp_clusters.mcmc}
\title{Most Probable Clusters}
\usage{
mp_clusters(x, ...)

\method{mp_clusters}{matrix}(x, rec_ids = NULL, ...)

\method{mp_clusters}{ExchangERResult}(x, ...)

\method{mp_clusters}{mcmc}(x, rec_ids = NULL, ...)
}
\arguments{
\item{x}{an \R object containing samples of clusterings. Currently, methods
are defined for \code{\linkS4class{ExchangERResult}} objects, \code{\link[coda:mcmc]{coda::mcmc}} objects
and \code{\link[base:matrix]{base::matrix}} objects (where rows index samples and columns index
cluster assignments for each record).}

\item{...}{further arguments passed to or from other methods.}

\item{rec_ids}{an optional vector of record identifiers. If NULL, the
record identifiers are extracted from \code{x} if available.  Otherwise
integers are used to identify the records.}
}
\value{
Returns a \code{MPClusters} object, which is a list containing the following
entries:
\item{clusters}{a list of vectors. The i-th entry specifies the most
frequently occurring cluster involving the i-th record among the samples
of clusterings. The cluster is represented as a vector of record
identifiers.}
\item{probabilities}{a vector of probabilities. The i-th entry specifies
the probability (frequency of occurrence) of the most frequently
occurring cluster involving the i-th record.}
}
\description{
Given samples of clusterings, this function returns a list containing the
\emph{most probable cluster} for each record.
}
\note{
The collection of most probable clusters is not a valid clustering, as
records may appear in multiple clusters. To obtain a clustering that
one can apply the \code{\link{smp_clusters}} function to the output of
this function.

In Steorts et al. (2016), the most probable clusters are referred to as
\emph{most probable maximal matching sets}.
}
\references{
Steorts, R. C., Hall, R. & Fienberg, S. E. A Bayesian Approach
to Graphical Record Linkage and Deduplication. \emph{JASA} \strong{111},
1660–1672 (2016).
}
