% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBDRP.R
\name{NBDRP}
\alias{NBDRP}
\title{NBD Random Partition}
\usage{
NBDRP(p, n, alpha, a, q)
}
\arguments{
\item{p}{Probability of success for the shifted negative binomial base
distribution on the cluster sizes. Must be a numeric scalar on the unit
interval \eqn{(0, 1]} or a \code{\link{BetaRV}} object.}

\item{n}{Target for number of successful trials for the shifted negative
binomial base distribution on the cluster sizes. Must be a positive
numeric scalar or a \code{\link{GammaRV}} object.}

\item{alpha}{Concentration parameter for the Dirichlet process prior on
the cluster sizes. Must be a positive numeric scalar.}

\item{a}{Target for number of successful trials for the shifted negative
binomial prior on the number of clusters. Must be a positive numeric
scalar.}

\item{q}{Probability of success for the shifted negative binomial prior on
the number of clusters. Must be a numeric scalar on the unit interval.}
}
\value{
The constructor \code{NBDRP} returns a \code{NBDRP} object which is a
subclass of \code{\linkS4class{RP}}.
}
\description{
TODO: update description.
A Kolchin partition model where the number of clusters is drawn from a
shifted negative binomial distribution, and the cluster sizes are drawn from
a distribution which is itself drawn from a Dirichlet process prior with
concentration \code{alpha} and a shifted negative binomial base distribution.
Following the paramterization of the negative binomial distribution in
\code{\link[stats:NegBinomial]{stats::NegBinomial}}:
\itemize{
\item the shifted negative binomial distribution on the number of clusters has
parameters \eqn{`size` = n} and \eqn{`prob` = p}.
\item the shifted negative binomial distribution for the cluster sizes has
parameters \eqn{`size` = a} and \eqn{`prob` = q}.
Hyperpriors on the \code{p} and \code{n} parameters are supported.
}
}
\references{
Brenda Betancourt, Giacomo Zanella, Hannah Wallach, Jeffrey W. Miller, Abbas
Zaidi & Rebecca C. Steorts (2016): Flexible Models for Microclustering with
Application to Entity Resolution, \emph{Advances in Neural Information Processing Systems}, 29
URL: https://proceedings.neurips.cc/paper_files/paper/2016/file/670e8a43b246801ca1eaca97b3e19189-Paper.pdf
}
\seealso{
Other random partitions defined in this package include \code{\link{PitmanYorRP}} and
\code{\link{EwensRP}}.
}
