% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PitmanYorRP.R
\name{PitmanYorRP}
\alias{PitmanYorRP}
\title{Pitman-Yor Random Partition}
\usage{
PitmanYorRP(alpha, d)
}
\arguments{
\item{alpha}{Concentration parameter that satisfies \eqn{alpha > -d}
or a \code{\link{GammaRV}}.}

\item{d}{Discount parameter that satisfies \eqn{0 \leq d < 1}{0 <= d < 1}
or a \code{\link{BetaRV}}.}
}
\value{
The constructor \code{PitmanYorRP} returns a \code{PitmanYorRP}
object which is a subclass of \code{\linkS4class{RP}}.
}
\description{
Represents a Pitman-Yor random partition.
}
\details{
Pitman-Yor random partitions are a subset of Ewens-Pitman random partitions
for which \eqn{0 \leq \sigma \leq 1}{0 <= σ <= 1} and
\eqn{\alpha > - \sigma}{ɑ > -σ}. These partitions can be generated by
sampling with replacement from an infinite population, where the mixing
proportions are drawn from a two-parameter Poisson-Dirichlet distribution.
}
\seealso{
Other random partitions defined in this package include \code{\link{EwensRP}} and
\code{\link{GeneralizedCouponRP}}.
}
