% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP.R
\docType{class}
\name{RP-class}
\alias{RP-class}
\title{Random partition}
\description{
A virtual class for a random partition, which is a subclass of
\code{\linkS4class{RV}}.
}
\details{
Three types of exchangeable random partitions are currently supported.
\itemize{
\item \emph{Pitman-Yor partition.} This yields a random partition where the
number of blocks scales sub-linearly (as a power law) in the
number of items. \code{\link{PitmanYorRP}} specifies a random partition of this
type.
\item \emph{Ewens partition.} This yields a random partition where the number
of blocks scales logarithmically in the number of items. \code{\link{EwensRP}}
specifies a random partition of this type.
\item \emph{Generalized coupon partition.} This yields a random partition where
the number of blocks approaches a constant asymptotically.
\code{\link{GeneralizedCouponRP}} specifies a random partition of this type.
}
}
