% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP.R, R/GeneralizedCouponRP.R, R/PitmanYorRP.R
\name{rrp}
\alias{rrp}
\alias{rrp,GeneralizedCouponRP-method}
\alias{rrp,PitmanYorRP-method}
\title{Generate a Random Partition}
\usage{
rrp(rp, n, ...)

\S4method{rrp}{GeneralizedCouponRP}(rp, n, ...)

\S4method{rrp}{PitmanYorRP}(rp, n, ...)
}
\arguments{
\item{rp}{a \code{\linkS4class{RP}} object}

\item{n}{number of integers in partition}

\item{...}{other parameters passed on to other methods.}
}
\value{
Returns a partition of the integers \eqn{\{1, 2, ..., n\}}{{1, 2, ..., n}}
as a list of blocks.
}
\description{
Generate a Random Partition
}
\section{Methods (by class)}{
\itemize{
\item \code{rrp(GeneralizedCouponRP)}: Specialization for \code{\link{GeneralizedCouponRV}}

\item \code{rrp(PitmanYorRP)}: Specialization for \code{\link{PitmanYorRP}}

}}
\examples{
rp <- GeneralizedCouponRP(1, 1)
rrp(rp, 100)

}
