% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RV.R, R/GammaRV.R, R/BetaRV.R, R/DirichletRV.R,
%   R/ShiftedNegBinomRV.R
\name{rrv}
\alias{rrv}
\alias{rrv,GammaRV-method}
\alias{rrv,BetaRV-method}
\alias{rrv,DirichletRV-method}
\alias{rrv,ShiftedNegBinomRV-method}
\title{Generate a Random Variable}
\usage{
rrv(rv, ...)

\S4method{rrv}{GammaRV}(rv, ...)

\S4method{rrv}{BetaRV}(rv, ...)

\S4method{rrv}{DirichletRV}(rv, ...)

\S4method{rrv}{ShiftedNegBinomRV}(rv, ...)
}
\arguments{
\item{rv}{a \code{\linkS4class{RV}} object}

\item{...}{other parameters passed on to other methods.}
}
\value{
Returns the generated value of the random variable.
}
\description{
Generate a Random Variable
}
\section{Methods (by class)}{
\itemize{
\item \code{rrv(GammaRV)}: Specialization for \code{\link{GammaRV}}

\item \code{rrv(BetaRV)}: Specialization for \code{\link{BetaRV}}

\item \code{rrv(DirichletRV)}: Specialization for \code{\link{DirichletRV}}

\item \code{rrv(ShiftedNegBinomRV)}: Specialization for \code{\link{ShiftedNegBinomRV}}

}}
\examples{
x <- BetaRV(1, 1)
rrv(x)

}
