% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_inference.R
\name{run_inference}
\alias{run_inference}
\alias{run_inference,ExchangERModel-method}
\alias{run_inference,ExchangERFit-method}
\title{Run Inference for a Model}
\usage{
run_inference(x, n_samples, thin_interval = 1, burnin_interval = 0, ...)

\S4method{run_inference}{ExchangERModel}(x, n_samples, thin_interval = 1, burnin_interval = 0, ...)

\S4method{run_inference}{ExchangERFit}(x, n_samples, thin_interval = 1, burnin_interval = 0, ...)
}
\arguments{
\item{x}{an \R object representing a model or a model with sampling
history. Currently, there are methods defined for \code{\linkS4class{ExchangERModel}}
objects and \code{\linkS4class{ExchangERFit}} objects. Passing a model begins
sampling from scratch, in which case a burn-in period is recommended.
Passing a model with sampling history will resume inference, adding
samples to the existing history.}

\item{n_samples}{a positive integer. The number of samples to generate
after applying burn-in and thinning.}

\item{thin_interval}{a positive integer. The period for saving samples
(intermediate samples are discarded). The default value is 1, which
means no thinning is applied.}

\item{burnin_interval}{a non-negative integer. The number of initial samples
to discard as burn-in. The default value is 0, which means no burn-in is
applied.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Returns an \code{\linkS4class{ExchangERFit}} object with the following slots:
\item{history}{a list containing the sampling history for
parameters/summary statistics along the Markov chain.}
\item{state}{a model object of the same class as the input, which
represents the state of all parameters in the last step of the Markov
chain.}
}
\description{
Runs approximate inference for a model using Markov chain Monte
Carlo (MCMC).
}
\section{Functions}{
\itemize{
\item \code{run_inference(ExchangERModel)}: Specialization for \code{\linkS4class{ExchangERModel}}

\item \code{run_inference(ExchangERFit)}: Specialization for \code{\linkS4class{ExchangERFit}}

}}
\examples{
## Initialize a model for entity resolution of RLdata500
library(comparator)
distort_prior <- BetaRV(1,5)
clust_prior <- PitmanYorRP(alpha = GammaRV(1.0, 1.0), d = BetaRV(1.0, 1.0))
attr_params <- list(
  fname_c1 = CategoricalAttribute(distort_prior),
  lname_c1 = CategoricalAttribute(distort_prior),
  bd = CategoricalAttribute(distort_prior),
  by = CategoricalAttribute(distort_prior),
  bm = CategoricalAttribute(distort_prior)
)
model <- exchanger(RLdata500, attr_params, clust_prior)

## Run inference
fit <- run_inference(model, n_samples=100, burnin_interval=100)

}
