% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_ep.R
\name{sample_ewens_pitman}
\alias{sample_ewens_pitman}
\title{Sample Ewens-Pitman partition}
\usage{
sample_ewens_pitman(n, alpha, sigma)
}
\arguments{
\item{n}{Number of elements to partition.}

\item{alpha, sigma}{Ewens-Pitman parameters as described in Pitman (2006).}
}
\value{
A partition of the integers {1, ..., n} represented as a list of
blocks.
}
\description{
This function uses a Chinese Restaurant Process construction to
generate a sample.
}
