% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RV.R
\name{sd}
\alias{sd}
\title{Standard Deviation of a Random Variable}
\usage{
sd(x, ...)
}
\arguments{
\item{x}{a \code{\linkS4class{RV}} object}

\item{...}{other parameters passed on to other methods.}
}
\value{
Returns the mean of the random variable
}
\description{
Computes the standard deviation of a random variable
}
\examples{
x <- BetaRV(1, 1)
sd(x)

}
