% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiftedNegBinomRV.R
\name{ShiftedNegBinomRV}
\alias{ShiftedNegBinomRV}
\title{Shifted Negative Binomial Random Variable}
\usage{
ShiftedNegBinomRV(size, prob)
}
\arguments{
\item{size}{Target number of successes before experiment is stopped.}

\item{prob}{Probability of success in each trial. Must be on the unit
interval \eqn{(0, 1]}}
}
\value{
The \code{ShiftedNegBinomRV} constructor returns a \code{ShiftedNegBinomRV}
object, which is a subclass of \code{\linkS4class{RV}}.
}
\description{
Represents a random variable with a negative binomial distribution,
shifted by 1 so that the support is the positive integers.
}
\details{
The parameterization adopted here matches \code{\link[stats:NegBinomial]{stats::NegBinomial}}.
The distribution is over the number of failures which occur in a sequence
of Bernoulli trials, before a target number of successes occur, given that
the number of failures is at least one.
If \eqn{x} denotes the number of failures, \eqn{size = r} denotes the
number of success by and \eqn{prob = p} denotes the probability of success
in each trial, then the density of \eqn{x} is
\deqn{p(X = x) = \frac{\Gamma(x + r - 1)}{(x - 1)! \Gamma(r)} p^n (1 - p)^{x - 1}}{p(X = x) = (Γ(x + r - 1))/((x - 1)! Γ(r)) p^n (1 - p)^{x - 1}}
for \eqn{x = 1, 2, 3, ...}, \eqn{n > 0} and \eqn{0 < p \leq 1}{0 < p <= 1}.
}
\examples{
x <- ShiftedNegBinomRV(100, 0.1)
mean(x)

}
\seealso{
Other random variables defined in this package
include \code{\link{BetaRV}}, \code{\link{GammaRV}} and \code{\link{DirichletRV}}.
}
