% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{smp_clusters}
\alias{smp_clusters}
\alias{smp_clusters.MPClusters}
\alias{smp_clusters.matrix}
\alias{smp_clusters.ExchangERFit}
\alias{smp_clusters.mcmc}
\title{Shared Most Probable Clusters}
\usage{
smp_clusters(x, ...)

\method{smp_clusters}{MPClusters}(x, ...)

\method{smp_clusters}{matrix}(x, rec_ids = NULL, ...)

\method{smp_clusters}{ExchangERFit}(x, ...)

\method{smp_clusters}{mcmc}(x, rec_ids = NULL, ...)
}
\arguments{
\item{x}{an \R object containing samples of clusterings or a summary
of the most probable clusters. Currently, methods are defined for
\code{MPClusters} objects (as output by \code{\link{mp_clusters}}),
\code{\linkS4class{ExchangERFit}} objects, \code{\link[coda:mcmc]{coda::mcmc}} objects
and \code{\link[base:matrix]{base::matrix}} objects (where rows index samples and columns index
cluster assignments for each record).}

\item{...}{further arguments passed to or from other methods.}

\item{rec_ids}{an optional vector of record identifiers. This argument is
ignored when an \code{MPClusters} object is passed for \code{x}. If NULL, the
record identifiers are extracted from \code{x} if available.  Otherwise
integers are used to identify the records.}
}
\value{
Returns the shared most probable clustering, represented as a list
of vectors. Each vector contains the identifiers of records belonging to
a cluster.
}
\description{
Computes a point estimate of the most likely clustering that obeys
transitivity constraints based on samples of clusterings. The method was
proposed by Steorts et al. (2016), where it is referred to as the
method of \emph{shared most probable maximal matching sets}.
}
\references{
Steorts, R. C., Hall, R. & Fienberg, S. E. A Bayesian Approach
to Graphical Record Linkage and Deduplication. \emph{JASA} \strong{111},
1660–1672 (2016).
}
\seealso{
The \code{\link[clevr:clustering_representations]{clevr::clusters_to_pairs}} and \code{\link[clevr:clustering_representations]{clevr::clusters_to_membership}}
functions are useful for converting the output of this function to different
clustering representations.
}
