% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_dist_fn.R
\name{transform_dist_fn}
\alias{transform_dist_fn}
\title{Transformed Distance Function}
\usage{
transform_dist_fn(dist_fn, threshold, scaling_factor = 1)
}
\arguments{
\item{dist_fn}{A vectorized distance function that takes a pair of
vectors y, x and returns their elementwise distances.}

\item{threshold}{a non-negative numeric. This threshold is applied to
scaled distances, so that any distances above the threshold are set to
\code{Inf}.}

\item{scaling_factor}{a non-negative numeric. Raw distances are multiplied by
this scaling factor. Defaults to 1.0.}
}
\value{
a transformed distance function
}
\description{
Generates a transformed distance function where scaling and thresholding
are applied.
}
\details{
The transformed distance function returns \code{Inf} if
\eqn{s \times dist_fn(y, x) > \tau}{s * dist_fn(y, x) > τ} where
\eqn{s} is the \code{scaling_factor} and \eqn{\tau}{τ} is the \code{threshold}.
Otherwise, the transformed distance function returns
\eqn{s \times dist_fn(y, x)}{s * dist_fn(y, x)}.
}
\seealso{
This is useful for preparing distance functions for \code{\link{Attribute}}.
}
