% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniformRP.R
\docType{class}
\name{UniformRP-class}
\alias{UniformRP-class}
\alias{UniformRP}
\title{Uniform Partition}
\value{
A \code{UniformRP} object which is a sub-class of
\code{\linkS4class{RP}}.
}
\description{
Specifies a uniform prior for the clustering---i.e. every possible
partitioning of the records into clusters is equally likely.
}
\seealso{
\code{\link{PitmanYorRP}},
\code{\link{EwensRP}}, \code{\link{GeneralizedCouponRP}}
}
