% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R, R/hello.R
\name{usd_rate}
\alias{usd_rate}
\title{Convert selected currencies via API}
\usage{
usd_rate(value, input = "EUR", output = "USD")

usd_rate(value, input = "EUR", output = "USD")
}
\arguments{
\item{value}{numeric vector}

\item{input}{character string}

\item{output}{character string}
}
\value{
result

result
}
\description{
Convert selected currencies via API

Convert selected currencies via API
}
