% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESCNBRP.R
\name{ESCNBRP}
\alias{ESCNBRP}
\title{ESC-NB Random Partition}
\usage{
ESCNBRP(p, n)
}
\arguments{
\item{p}{Probability of success. Must be a numeric scalar on the unit
interval \eqn{(0, 1]} or a \code{\link{BetaRV}} object.}

\item{n}{Target for number of successful trials, or dispersion parameter
(the shape parameter of the gamma mixing distribution). Must be a positive
numeric scalar or a \code{\link{GammaRV}} object.}
}
\value{
The constructor \code{ESCNBRP} returns a \code{ESCNBRP} object which is a
subclass of \code{\linkS4class{RP}}.
}
\description{
Represents an ESC-NB (Exchangeable Sequences of Clusters-Negative Binomial)
random partition.
}
\details{
An ESC model where the cluster sizes are drawn from a shifted negative
binomial distribution (with support on the positive integers).
The parameterization of the shifted negative binomial distribution matches
\code{\link[stats:NegBinomial]{stats::NegBinomial}}.
For \eqn{`size` = n} and \eqn{`prob` = p}, the probability mass associated
with cluster size \eqn{x} is:
\deqn{p(X = x) = \frac{\Gamma(x + r - 1)}{(x - 1)! \Gamma(r)} p^n (1 - p)^{x - 1}}{p(X = x) = (Γ(x + r - 1))/((x - 1)! Γ(r)) p^n (1 - p)^{x - 1}}
for \eqn{x = 1, 2, 3, ...}, \eqn{n > 0} and \eqn{0 < p \leq 1}{0 < p <= 1}.
Hyperpriors on the \code{p} and \code{n} parameters are supported.
}
\references{
Brenda Betancourt, Giacomo Zanella & Rebecca C. Steorts (2020): Random
Partition Models for Microclustering Tasks, \emph{Journal of the American
Statistical Association}, DOI: 10.1080/01621459.2020.1841647.
}
\seealso{
Other random partitions defined in this package include \code{\link{PitmanYorRP}} and
\code{\link{EwensRP}}.
}
