% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R, R/hello.R
\name{uneur}
\alias{uneur}
\title{EUR string to numeric converter}
\usage{
uneur(str)

uneur(str)
}
\arguments{
\item{str}{character string}
}
\value{
EUR in numeric format

EUR in numeric format
}
\description{
EUR string to numeric converter

EUR string to numeric converter
}
