% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RV.R, R/GammaRV.R, R/BetaRV.R, R/DirichletRV.R,
%   R/ShiftedNegBinomRV.R
\name{var}
\alias{var}
\alias{var,GammaRV-method}
\alias{var,BetaRV-method}
\alias{var,DirichletRV-method}
\alias{var,ShiftedNegBinomRV-method}
\title{Variance-Covariance of a Random Variable}
\usage{
var(x, ...)

\S4method{var}{GammaRV}(x)

\S4method{var}{BetaRV}(x)

\S4method{var}{DirichletRV}(x)

\S4method{var}{ShiftedNegBinomRV}(x)
}
\arguments{
\item{x}{a \code{\linkS4class{RV}} object}

\item{...}{other parameters passed on to other methods.}
}
\value{
Returns the variance if the random variable is a scalar, or a
variance-covariance matrix if the random variable is a vector.
}
\description{
Computes the variance-covariance matrix of a random variable
}
\section{Methods (by class)}{
\itemize{
\item \code{var(GammaRV)}: Specialization for \code{\link{GammaRV}}

\item \code{var(BetaRV)}: Specialization for \code{\link{BetaRV}}

\item \code{var(DirichletRV)}: Specialization for \code{\link{DirichletRV}}

\item \code{var(ShiftedNegBinomRV)}: Specialization for \code{\link{ShiftedNegBinomRV}}

}}
\examples{
x <- BetaRV(1, 1)
var(x)

}
